requirements.txt
pytest
playwright
allure-pytest
requests
pandas
appium-python-client
selenium
jira